refactor : `image-carousel` -> `image-slider` 변경
